requirements.txt
absl-py
opencv-python-headless
update
Click
Flask
gast
google-pasta
grpcio
gunicorn
h5py
idna
itsdangerous
Jinja2
Keras
Markdown
MarkupSafe
numpy
Pillow
protobuf
python-resize-image
requests
six
tensorflow-cpu
urllib3
uWSGI
Werkzeug
wrapt